update Cookie class to store the state of the checkbox filter
State is maintained only while the popup is open.

Closing and reopening the popup causes all checkboxed to revert
to the default state, which is checked to allow the cookie to export.

Also worth noting is that perfomance should be greatly improved.
Nearly every button click was triggering a reload of all cookies,
which caused a loss of state. Now, this is done selectively.
When cookies haven't changed, only the DOM is updated.